multi: comment out to-be-used code - also remove unnecessary conversion

+++ ui/decredmaterial/button.go
@@ -154,12 +154,12 @@ func drawInk(gtx *layout.Context, c widget.Click) {
 	}).Add(gtx.Ops)
 	clip.Rect{
 		Rect: f32.Rectangle{Max: f32.Point{
-			X: float32(size),
-			Y: float32(size),
+			X: size,
+			Y: size,
 		}},
 		NE: rr, NW: rr, SE: rr, SW: rr,
 	}.Op(gtx.Ops).Add(gtx.Ops)
-	paint.PaintOp{Rect: f32.Rectangle{Max: f32.Point{X: float32(size), Y: float32(size)}}}.Add(gtx.Ops)
+	paint.PaintOp{Rect: f32.Rectangle{Max: f32.Point{X: size, Y: size}}}.Add(gtx.Ops)
 	stack.Pop()
 	op.InvalidateOp{}.Add(gtx.Ops)
 }
